refactor: move db as private repo of server
