+++ ee/tabby-webserver/src/schema/auth.rs
